+++ packages/drift_sqlite_async/test/utils/test_utils.dart
@@ -1,31 +1,11 @@
 import 'dart:io';
 
-import 'package:glob/glob.dart';
-import 'package:glob/list_local_fs.dart';
-import 'package:sqlite_async/sqlite3_common.dart';
 import 'package:sqlite_async/sqlite_async.dart';
 import 'package:test_api/src/backend/invoker.dart';
 
-class TestSqliteOpenFactory extends DefaultSqliteOpenFactory {
-  TestSqliteOpenFactory({
-    required super.path,
-    super.sqliteOptions,
-  });
-
-  @override
-  CommonDatabase open(SqliteOpenOptions options) {
-    final db = super.open(options);
-
-    return db;
-  }
-}
-
-DefaultSqliteOpenFactory testFactory({String? path}) {
-  return TestSqliteOpenFactory(path: path ?? dbPath());
-}
-
 Future<SqliteDatabase> setupDatabase({String? path}) async {
-  final db = SqliteDatabase.withFactory(testFactory(path: path));
+  final db =
+      SqliteDatabase.withFactory(SqliteOpenFactory(path: path ?? dbPath()));
   await db.initialize();
   return db;
 }
@@ -46,15 +26,6 @@ Future<void> cleanDb({required String path}) async {
   } on PathNotFoundException {
     // Not an issue
   }
-}
-
-List<String> findSqliteLibraries() {
-  var glob = Glob('sqlite-*/.libs/libsqlite3.so');
-  List<String> sqlites = [
-    'libsqlite3.so.0',
-    for (var sqlite in glob.listSync()) sqlite.path
-  ];
-  return sqlites;
 }
 
 String dbPath() {